requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
python-binance>=1.0.19
pandas>=2.2.2
numpy>=1.26.4

# ── ML ────────────────────────────────────────────────────────────────────────
xgboost>=2.0.3
scikit-learn>=1.5.0
joblib>=1.4.2

# ── Technical Analysis ────────────────────────────────────────────────────────
ta>=0.11.0
matplotlib>=3.8.0

# ── API / Web ─────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn>=0.29.0
requests>=2.31.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30

# ── Monitoring ────────────────────────────────────────────────────────────────
psutil>=5.9.8

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
tabulate>=0.9.0
python-dateutil>=2.9.0